✨ 구글 소셜 로그인 스켈레톤 구현
