+++ package.json
@@ -16,6 +16,7 @@
     "check-all": "yarn types:check && yarn format:check && yarn lint:check && yarn unused:check"
   },
   "dependencies": {
+    "@react-oauth/google": "0.12.1",
     "@tanstack/react-query": "5.62.11",
     "react": "18.3.1",
     "react-dom": "18.3.1",
