+++ src/entities/route.ts
@@ -2,4 +2,5 @@ export const PATH = {
   INDEX: '/',
   ECHO: '/echo',
   SIGN_UP: '/signup',
+  VERIFY_EMAIL: '/verify-email',
 };
