+++ src/main.tsx
@@ -12,6 +12,7 @@ if (root === null) throw new Error('Root element not found');
 const ENV = {
   APP_ENV: import.meta.env.MODE as 'prod' | 'dev' | 'mock',
   API_BASE_URL: import.meta.env.VITE_API_BASE_URL,
+  GOOGLE_CLIENT_ID: import.meta.env.VITE_GOOGLE_CLIENT_ID,
 };
 
 async function enableMocking() {
