+++ src/pages/EmailVerifyPage/index.tsx
@@ -1,0 +1,3 @@
+export const EmailVerifyPage = () => {
+  return <div>이메일 인증 페이지입니다.</div>;
+};
